# Improve perf of Enumerable.Sum/Average/Max/Min for arrays and lists