… src/libraries/System.Linq/src/System/Linq/Sum.cs
… src/libraries/System.Linq/tests/AverageTests.cs
… src/libraries/System.Linq/tests/MaxTests.cs
… src/libraries/System.Linq/tests/MinTests.cs
… src/libraries/System.Linq/tests/Shuffler.cs