@eiriktsarpalis 2022-02-01
> @eiriktsarpalis, please let me know in general if you're ok with the extra code here for this special-casing.

Should be ok, but per your own comment in https://github.com/dotnet/runtime/pull/64470#discussion_r794936082 it would be nice if the code be moved out of Linq eventually.