@tannergooding 2022-02-01
> I assume the use of vectorization for floats/doubles could change the answer in some cases due to lack of associativity, yes? Thoughts on how much we should care about that? Also, it seemed like it should be possible to vectorize some of the methods doing checked arithmetic, but I wasn't sure how to do so correctly and skipped those. It also seemed like it should be possible to vectorize min/max for floats/doubles, but they have special-handling of NaN I couldn't figure out how to replicate with Vector.

@stephentoub right. There are a number of cases where we cannot trivially vectorize float/double (particularly operations that compute a new value like `Sum`) because it can drastically change the output and will lead to determinism issues among other bugs.

The simplest scenario to explain is that with `integers` the delta between one representable value and the next is always `1` (ignoring overflow for a minute). Where-as with `floating-point` this starts at `epsilon` (the smallest representable value greater than zero) and then doubles every power of `2` after that. For `float` this means that for `2^22` to `2^23` the delta between values is 0.5, for `2^23` to `2^24` the delta between values is `1`, for `2^24` to 2^25 it is `2`, at 2^26 it is `4`, and so on (in both directions going down, including to negative powers until you hit `epsilon` and going up until you hit `float.MaxValue`).

What this means is that if you take a case like `new float[] { 16777216.0f, 0.5f, 0.5f, 0.5f }` you will get back `16777216.0f` because `2^24 + 0.5` always returns `2^24` since its less than the delta between values (`2`). However if you reorder this to `new float[] { 0.5f, 0.5f, 0.5f, 16777216.0f }` you instead get back `16777218.0f`, because `0.5 + 0.5 +0.5` is enough to create `1.5` and then (due to rounding) is enough to round up to 16777218.

Vectorization is impacted here because it ends up adding `n` to `n + Count` and then adding them "across" at the end. There have been open asks for several years to provide a "fast math" switch that would allow such optimizations but its non-trivial to achieve. The easiest thing would be to expose a new overload that allows users to explicitly opt-into such differences.

------------------

Things that do not compute a "new value" like Min/Max (it only does a comparison and returns one of the inputs) can be vectorized but it needs to take into account things like NaN handling as you indicated.