@stephentoub 2022-02-03
>> I assume the use of vectorization for floats/doubles could change the answer in some cases due to lack of associativity, yes? Thoughts on how much we should care about that?

> I think we should care a lot and not do this for float/double

Ok, deleting.